NTP-1218: Test fix for polling

+++ Application/Common/Interfaces/Repositories/IGenericRepository.cs
@@ -4,7 +4,7 @@ namespace Application.Common.Interfaces.Repositories;
 
 public interface IGenericRepository<T> where T : class
 {
-    T GetById(long id);
+    T GetById(int id);
     IEnumerable<T> GetAll();
 
     //includeProperties - i.e: "MagicLink,TuitionPartner"
